requirements.txt
asgiref==3.4.1
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.7
colorama==0.4.4
Django==3.2.8
django-extensions==3.1.3
django-jazzmin==2.4.8
django-js-asset==1.2.2
django-mptt==0.13.4
django-ninja==0.16.1
django-silk==4.2.0
django-simple-history==3.0.0
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
gprof2dot==2021.2.21
idna==3.3
install==1.3.4
Jinja2==3.0.2
MarkupSafe==2.0.1
Pillow==8.4.0
pretty-errors==1.2.24
pyasn1==0.4.8
pycodestyle==2.8.0
pydantic==1.8.2
pydot==1.4.2
Pygments==2.10.0
pyparsing==3.0.4
python-dateutil==2.8.2
python-jose==3.3.0
pytz==2021.3
requests==2.26.0
rsa==4.7.2
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.2
